move parsing logic in md translator
