+++ cs/Markdown/Tokens/Token.cs
@@ -7,13 +7,15 @@
         public int Length => Value.Length;
         public int Position { get; }
         public bool IsOpening { get; }
+        public bool ShouldBeSkipped { get; }
 
-        public Token(string value, TokenType type, int position, bool isOpening)
+        public Token(string value, TokenType type, int position, bool isOpening, bool shouldBeSkipped)
         {
             Value = value;
             Type = type;
             Position = position;
             IsOpening = isOpening;
+            ShouldBeSkipped = shouldBeSkipped;
         }
     }
 }
